[SPARK-55910][SQL][TESTS][FOLLOWUP] Move compareAnswers into object QueryTest
### What changes were proposed in this pull request?

Followup to SPARK-55910. Move the `compareAnswers` helper from `object SQLTestUtils` into `object QueryTest`. After the move `object SQLTestUtils` is empty and is removed; the two existing callers in `SparkPlanTest.scala` are updated to call `QueryTest.compareAnswers`.

### Why are the changes needed?

Following the SQLTestUtilsBase/SQLTestUtils -> QueryTestBase/QueryTest trait consolidation in SPARK-55910, the only remaining piece in the SQLTestUtils file was `object SQLTestUtils.compareAnswers`. Moving it to `object QueryTest` finishes the consolidation and removes the last reason to reference `object SQLTestUtils`.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Existing `SparkPlanTest`-based suites exercise both call sites.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Claude Opus 4.7
