+++ sql/core/src/test/scala/org/apache/spark/sql/test/SQLTestUtils.scala
@@ -19,8 +19,7 @@ package org.apache.spark.sql.test
 
 import org.scalatest.Suite
 
-import org.apache.spark.sql.{QueryTest, QueryTestBase, Row}
-import org.apache.spark.sql.catalyst.util._
+import org.apache.spark.sql.{QueryTest, QueryTestBase}
 
 /**
  * Kept as an empty alias of [[QueryTest]] for backward compatibility with existing subclasses.
@@ -33,46 +32,3 @@ private[sql] trait SQLTestUtils extends QueryTest
  * New test suites should extend [[QueryTestBase]] directly.
  */
 private[sql] trait SQLTestUtilsBase extends QueryTestBase { self: Suite => }
-
-private[sql] object SQLTestUtils {
-
-  def compareAnswers(
-      sparkAnswer: Seq[Row],
-      expectedAnswer: Seq[Row],
-      sort: Boolean): Option[String] = {
-    def prepareAnswer(answer: Seq[Row]): Seq[Row] = {
-      // Converts data to types that we can do equality comparison using Scala collections.
-      // For BigDecimal type, the Scala type has a better definition of equality test (similar to
-      // Java's java.math.BigDecimal.compareTo).
-      // For binary arrays, we convert it to Seq to avoid of calling java.util.Arrays.equals for
-      // equality test.
-      // This function is copied from Catalyst's QueryTest
-      val converted: Seq[Row] = answer.map { s =>
-        Row.fromSeq(s.toSeq.map {
-          case d: java.math.BigDecimal => BigDecimal(d)
-          case b: Array[Byte] => b.toSeq
-          case o => o
-        })
-      }
-      if (sort) {
-        converted.sortBy(_.toString())
-      } else {
-        converted
-      }
-    }
-    if (prepareAnswer(expectedAnswer) != prepareAnswer(sparkAnswer)) {
-      val errorMessage =
-        s"""
-           | == Results ==
-           | ${sideBySide(
-          s"== Expected Answer - ${expectedAnswer.size} ==" +:
-            prepareAnswer(expectedAnswer).map(_.toString()),
-          s"== Actual Answer - ${sparkAnswer.size} ==" +:
-            prepareAnswer(sparkAnswer).map(_.toString())).mkString("\n")}
-      """.stripMargin
-      Some(errorMessage)
-    } else {
-      None
-    }
-  }
-}
